+++ docs/src/Fulma/Components/Dropdown/Types.fs
@@ -1,0 +1,8 @@
+module Components.Dropdown.Types
+
+type Model =
+    { Intro : string
+      BasicViewer : Viewer.Types.Model }
+
+type Msg =
+    | BasicViewerMsg of Viewer.Types.Msg